requirements.txt
django==1.11
djangorestframework==3.6.2
jsonfield==2.0.1

